requirements.txt
# Sales Insights AI - Professional Dependencies
# Developed by: João Gabriel de Araujo Diniz

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & ORM
sqlalchemy==2.0.23
sqlite3

# AI & Machine Learning
langchain==0.0.350
openai==0.28.1

# Data Processing
pandas==2.1.4
numpy==1.24.3

# API & Validation
pydantic==2.5.0
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Security
cryptography==41.0.8

# Utilities
requests==2.31.0
python-dateutil==2.8.2

